Create README.md
This is Repository with Assignments and project of "Pubg game strategy". This project deals with the PUBG game plan to get victory. By analyzing the project, we can get best
strategy to win like “Hiding in One Spot” or “TO be the Top Shot”. In this project we can analyze
the data with the help of Plots.